Separated classes so main.py is more readable
