+++ high_scores.txt
@@ -100,3 +100,17 @@
 0
 2
 4
+2
+1
+2
+3
+3
+81
+29
+0
+0
+0
+37
+23
+4
+101
