+++ platform.py
@@ -1,0 +1,10 @@
+import pygame
+from constants import *
+
+# Define the platform class
+class Platform(pygame.sprite.Sprite):
+    def __init__(self, width, height):
+        super().__init__()
+        self.image = pygame.Surface([width, height])
+        self.image.fill(GREEN)
+        self.rect = self.image.get_rect()
